requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
validators==0.20.0
python-dotenv==0.19.0
requests==2.28.1
